move metadata-header into own vue file
